:+1: go-test: badrequestのケース追加
