+++ test/go_test/main.go
@@ -7,4 +7,5 @@ import (
 func main() {
 	test_cases.Get()
 	test_cases.Error()
+	test_cases.BadRequest()
 }
