+++ test/go_test/test_cases/get.go
@@ -1,107 +1,11 @@
 package test_cases
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
 )
-
-const (
-	baseURL = "http://127.0.0.1"
-
-	red   = "\033[31m"
-	green = "\033[32m"
-	blue  = "\033[34m"
-	reset = "\033[39m"
-)
-
-// 長さ1001の文字列生成
-func string1001() string {
-	var s string
-	for i := 0; i < 1001; i++ {
-		s += "a"
-	}
-	return s
-}
-
-// エラーケース
-func Error() {
-	log.Printf("%sRUN Error Test%s", blue, reset)
-	defer log.Printf("%sEND Error Test%s", blue, reset)
-	tests := []struct {
-		name         string
-		request      *http.Request
-		expectStatus int
-	}{
-		{
-			name: "NotFound",
-			request: func() *http.Request {
-				port := ":1111"
-				target := "/no_such_file"
-				url := baseURL + port + target
-				req, err := http.NewRequest("GET", url, nil)
-				if err != nil {
-					panic(fmt.Errorf("NewRequest: %w", err))
-				}
-				return req
-			}(),
-			expectStatus: http.StatusNotFound,
-		},
-		{
-			name: "MethodNotAllowed",
-			request: func() *http.Request {
-				port := ":1111"
-				target := ""
-				url := baseURL + port + target
-				req, err := http.NewRequest("HOGE", url, nil)
-				if err != nil {
-					panic(fmt.Errorf("NewRequest: %w", err))
-				}
-				return req
-			}(),
-			expectStatus: http.StatusMethodNotAllowed,
-		},
-		{
-			name: "RequestEntityTooLarge",
-			request: func() *http.Request {
-				port := ":1111"
-				target := ""
-				url := baseURL + port + target
-				body := bytes.NewBufferString(string1001())
-				req, err := http.NewRequest("POST", url, body)
-				if err != nil {
-					panic(fmt.Errorf("NewRequest: %w", err))
-				}
-				return req
-			}(),
-			expectStatus: http.StatusRequestEntityTooLarge,
-		},
-	}
-
-	client := http.DefaultClient
-	for _, tt := range tests {
-		log.Println(" Run:", tt.name)
-
-		resp, err := client.Do(tt.request)
-		if err != nil {
-			log.Println(red, " [ERROR]", reset, err)
-			continue
-		}
-		defer resp.Body.Close()
-
-		if resp.StatusCode != tt.expectStatus {
-			log.Printf(" %s[ERROR]%s status code\n\texpect: %d\n\tactual: %d",
-				red,
-				reset,
-				tt.expectStatus,
-				resp.StatusCode)
-			continue
-		}
-		log.Println(green, "OK", reset)
-	}
-}
 
 func Get() {
 	log.Printf("%sRUN Get Test%s", blue, reset)
